refactor: moved all components into respective component folders
