+++ src/components/EntryEditor/EmbedImageDialog.js
