+++ src/components/EntryEditor/PropertyEditor.js
