+++ src/components/EntryEditor/PropertySelector.js
