+++ src/components/LogDetails/GroupHeader.js
