+++ src/components/LogDetails/LogDetails.js
@@ -19,10 +19,10 @@
 import React, { Component } from 'react';
 import Container from 'react-bootstrap/Container';
 import { Remarkable } from 'remarkable';
-import imageProcessor from '../utils/image-processor';
-import '../css/olog.css';
-import customization from '../utils/customization';
-import {getLogEntryGroupId} from '../utils/utils';
+import imageProcessor from '../../utils/image-processor';
+import '../../css/olog.css';
+import customization from '../../utils/customization';
+import {getLogEntryGroupId} from '../../utils/utils';
 import ToggleButton from 'react-bootstrap/ToggleButton';
 import ButtonGroup from 'react-bootstrap/ButtonGroup';
 import Button from 'react-bootstrap/Button';
