+++ src/components/LogDetails/NavigationButtons.js
