+++ src/components/LogDetails/Properties.js
