+++ src/components/SearchResult/SearchResultDay.js
