+++ src/components/SearchResult/SearchStringHelpDialog.js
